% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABS.R
\name{ABS}
\alias{ABS}
\title{Compliles a tidy dataset from ABS data as read in using read_ABS}
\usage{
ABS(file, sheets, siccodes)
}
\arguments{
\item{file}{an ABS workbook}

\item{sheets}{a vector of spreadsheet tab names}

\item{siccodes}{a vector of SIC codes (as characters)}
}
\value{
a data frame
}
\description{
Uses read_ABS to compile and concatenate a dataset from multiple tabs in a
'standard' ABS workbook. The data is transformed to tidy (long) form.
}
\examples{
ABS("abs_data.xls", c("Section C"), c("10", "11") )
}
