% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital.R
\name{capital}
\alias{capital}
\title{Get capital data}
\usage{
capital(series, from, to)
}
\arguments{
\item{series}{a vector of ONS CDID identifiers for capital stocks}

\item{from}{a start year}

\item{to}{an end year}
}
\description{
Leverages read_ONS - it would be better to try and generalise this function
}
