% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflators.R
\name{deflators}
\alias{deflators}
\title{Get deflator data}
\usage{
deflators(series, from, to)
}
\arguments{
\item{series}{a vector of ONS CDID identifiers}

\item{from}{a start year}

\item{to}{an end year}
}
\description{
Leverages read_ONS- it would be better to try and generalise this function
}
