% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_weights.R
\name{input_weights}
\alias{input_weights}
\title{calculate input weights for a sector}
\usage{
input_weights(data, period, capital, purchases, employment)
}
\arguments{
\item{data}{a dataset containing current price data}

\item{period}{the period variable}

\item{capital}{the capital variable}

\item{purchases}{the purchases variable}

\item{employment}{the employment variable}
}
\value{
a dataset of input weights
}
\description{
calculate input weights for a sector
}
\examples{

}
