% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ons_series}
\alias{ons_series}
\title{ONS series data}
\format{An object of class \code{data.frame} with 130 rows and 4 columns.}
\usage{
ons_series
}
\description{
http://r-pkgs.had.co.nz/data.html
}
\keyword{datasets}
