% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ABS.R
\name{read_ABS}
\alias{read_ABS}
\title{reads in a sheet from a standard ONS ABS spreadsheet}
\usage{
read_ABS(file, sheet)
}
\arguments{
\item{file}{spreadsheet file name}

\item{sheet}{spreadsheet sheet name}
}
\value{
a data frame
}
\description{
Expects a 'standard' ABS spreadsheet as published at (link). Applies standard
column names and variable types, skips the first 8 header lines, and removes
blank lines. Fills down SIC codes into empty cells, but does not do this for
SIC descriptions, since some are displayed on multiple lines.
}
\examples{

}
