% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ASHE.R
\name{read_ASHE}
\alias{read_ASHE}
\title{Read in an ASHE data table}
\usage{
read_ASHE(file, sheet, year, category)
}
\arguments{
\item{file}{spreadsheet workbook containing ASHE data}

\item{sheet}{worksheet tab to process}

\item{year}{provide a year to add to the output dataset}

\item{category}{provide a category to add to the dataset}
}
\value{
a dataset of ASHE data
}
\description{
written for table 4.9, maybe works with other downloads
}
\examples{
read_ASHE("ASHE_T4_9.xls", "Full time", 2017, "fulltime")
}
