% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ONS.R
\name{read_ONS}
\alias{read_ONS}
\title{Extract an ONS data series}
\usage{
read_ONS(seriesuri, from, to)
}
\arguments{
\item{seriesuri}{the URI of the data series}

\item{from}{the start year}

\item{to}{the end year}
}
\value{
a dataframe of an ONS data series
}
\description{
creates a url which returns a yearly series of ONS data
}
\examples{
read_ONS("/economy/inflationandpriceindices/timeseries/l55o/mm23", 2005, 2016)
}
