% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sectors}
\alias{sectors}
\title{ONS sectors data}
\format{An object of class \code{data.frame} with 32 rows and 2 columns.}
\usage{
sectors
}
\description{
http://r-pkgs.had.co.nz/data.html
}
\keyword{datasets}
