% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFP.R
\docType{package}
\name{TFP}
\alias{TFP}
\alias{TFP-package}
\title{produce Total Factor Productivity of the food chain statistics relases}
\description{
The TFP package contains a number of functions for working with and
calculating TFP. Maybe one day it will be a RAP!
}
